requirements.txt
cffi==1.14.6
cryptography==3.4.8
pycparser==2.20
pycrypto==2.6.1
